requirements.txt
torch>=2.2.0
pandas>=2.1.0
numpy>=1.26.0
pyyaml>=6.0.1
ta>=0.10.2
pyarrow>=15.0.0
websockets>=11.0.3
httpx>=0.27.0
openai>=1.37.0
loguru>=0.7.2
tenacity>=8.2.3
